knowledge_base/requirements.txt
fastapi
uvicorn
huggingface-hub
InstructorEmbedding
faiss-cpu
llama-index>=0.11.0
llama-index-embeddings-huggingface
llama-index-llms-llama-cpp
llama-index-vector-stores-faiss
pdfplumber
pydantic
python-docx
python-multipart